[rs] Refactor parser modules
