+++ rs/src/complete/button.rs
@@ -1,7 +1,7 @@
-use crate::parsers::basic_data_types::{parse_color_transform_with_alpha, parse_matrix};
-use crate::parsers::display::{parse_blend_mode, parse_filter_list};
-use crate::parsers::sound::parse_sound_info;
-use nom::number::streaming::{le_u16 as parse_le_u16, le_u8 as parse_u8};
+use crate::complete::display::{parse_blend_mode, parse_filter_list};
+use crate::complete::sound::parse_sound_info;
+use crate::streaming::basic_data_types::{parse_color_transform_with_alpha, parse_matrix};
+use nom::number::complete::{le_u16 as parse_le_u16, le_u8 as parse_u8};
 use nom::IResult as NomResult;
 use swf_tree as ast;
