+++ rs/src/complete/mod.rs
@@ -1,0 +1,15 @@
+pub(crate) mod button;
+pub(crate) mod display;
+pub(crate) mod gradient;
+pub(crate) mod image;
+pub(crate) mod morph_shape;
+pub(crate) mod movie;
+pub(crate) mod shape;
+pub(crate) mod sound;
+pub(crate) mod tag;
+pub(crate) mod text;
+pub(crate) mod video;
+
+pub use movie::parse_swf;
+pub use movie::SwfParseError;
+pub use tag::parse_tag;
