+++ rs/src/complete/video.rs
@@ -1,4 +1,4 @@
-use nom::number::streaming::le_u8 as parse_u8;
+use nom::number::complete::le_u8 as parse_u8;
 use nom::IResult as NomResult;
 use swf_tree as ast;
